Move RTEngine traits to its own module
This is just for code clarification, as they are the reexported by the main traits.rs file.
